List, tuple, range and mutability
A list is a collection of objects separated by comma and enclosed in a square bracket []. They are mutable i.e there values can be edited and changed

A tuple is more like a string, once the values have been inputted they cannot be changed(they are immutable). A tuple is most times used to describe an object. E.G a student's ID

A range is often used to generate a sequence of numbers. It doesn’t store all the numbers in memory — it creates them as needed (which is super memory-efficient).

 Mutability refers to whether or not an object can be changed after it’s created.